+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Change
+- Allow relative values for `gravity` and `watermark` offsets.
 
 ## [3.21.0] - 2023-11-23
 ### Add